renames index to DownloadButton
